code/art-gallery-ai/requirements.txt
streamlit
langchain==0.1.0
openai
python-dotenv

